Move pytest fixtures from agent_manager_test.py to conftest.py
Problem: Three pytest fixtures (broadcaster, agent_manager, git_work_dir) were
defined directly in agent_manager_test.py, violating the CLAUDE.md rule that
all fixtures must be in conftest.py.
Fix: Created conftest.py in imbue/minds_workspace_server/ with the three fixtures.
The agent_manager fixture now uses monkeypatch.setenv (standard pytest) instead
of a custom _env context manager, avoiding the need to share that helper class.
